requirements.txt
aiogram==3.10.0
APScheduler==3.10.4
httpx==0.27.2
python-dotenv==1.0.1
pytz==2024.1
